requirements.txt
# AI CustomURL Extension for ComfyUI
# Core dependencies

requests>=2.31.0
aiohttp>=3.9.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
opencv-python>=4.8.0

# Optional but recommended
typing-extensions>=4.8.0

